moving a function
