+++ src/services/MatchService.ts
@@ -8,6 +8,7 @@ import ActiveMatchRepository from "../repositories/ActiveMatchRepository";
 import QueueRepository from "../repositories/QueueRepository";
 import { NewActiveMatchInput } from "../repositories/ActiveMatchRepository/types";
 import { PlayerInQueue } from "../repositories/QueueRepository/types";
+import { Team } from "../types/common";
 
 export async function createRandomMatch(): Promise<Array<NewActiveMatchInput>> {
   const ballChasers = await QueueRepository.getAllBallChasersInQueue();
@@ -38,4 +39,31 @@ export async function blueCaptainsChoice(chosenPlayers: string | null): Promise<
 
 export async function orangeCaptainsChoice(chosenPlayers: string[]): Promise<void> {
   await orangePlayerChosen(chosenPlayers);
+}
+
+export async function createMatchFromChosenTeams(): Promise<Array<NewActiveMatchInput>> {
+  const createdTeams: NewActiveMatchInput[] = [];
+
+  //Update the last available player to blue and push all players into Active Match
+  const ballchasers = await QueueRepository.getAllBallChasersInQueue();
+  for (const p of ballchasers) {
+    if (p.team !== null) {
+      createdTeams.push({ id: p.id, team: p.team });
+    } else {
+      await QueueRepository.updateBallChaserInQueue({
+        id: p.id,
+        team: Team.Blue,
+      });
+      createdTeams.push({ id: p.id, team: Team.Blue });
+    }
+  }
+
+  await Promise.all([
+    //Create Match
+    await ActiveMatchRepository.addActiveMatch(createdTeams),
+    //Match is created, reset the queue.
+    await QueueRepository.removeAllBallChasersFromQueue(),
+  ]);
+
+  return createdTeams;
 }
